requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
commonmark==0.9.1
idna==3.3
Pygments==2.12.0
python-frontmatter==1.0.0
PyYAML==5.1
requests==2.27.1
rich==12.5.1
urllib3==1.26.9
